Crate google_webmasters3

This documentation was generated from webmasters crate version 1.0.0+20160317, where 20160317 is the exact revision of the webmasters:v3 schema built by the mako code generator v1.0.0.

Everything else about the webmasters v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

searchanalytics

query

sitemaps

delete, get, list and submit

sites

add, delete, get and list

urlcrawlerrorscounts

query

urlcrawlerrorssamples

get, list and mark as fixed

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.sitemaps().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-webmasters3 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_webmasters3 as webmasters3;
use webmasters3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use webmasters3::Webmasters;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Webmasters::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.sitemaps().get("siteUrl", "feedpath")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

ApiDataRow
There is no detailed description.

ApiDimensionFilter
There is no detailed description.

ApiDimensionFilterGroup
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

SearchAnalyticsQueryRequest
There is no detailed description.

SearchAnalyticsQueryResponse
A list of rows, one per result, grouped by key. Metrics in each row are aggregated for all data grouped by that key either by page or property, as specified by the aggregation type parameter.

SearchanalyticMethods
A builder providing access to all methods supported on searchanalytic resources. It is not used directly, but through the Webmasters hub.

SearchanalyticQueryCall
Query your data with filters and parameters that you define. Returns zero or more rows grouped by the row keys that you define. You must define a date range of one or more days.

SiteAddCall
Adds a site to the set of the user's sites in Search Console.

SiteDeleteCall
Removes a site from the set of the user's Search Console sites.

SiteGetCall
Retrieves information about specific site.

SiteListCall
Lists the user's Search Console sites.

SiteMethods
A builder providing access to all methods supported on site resources. It is not used directly, but through the Webmasters hub.

SitemapDeleteCall
Deletes a sitemap from this site.

SitemapGetCall
Retrieves information about a specific sitemap.

SitemapListCall
Lists the sitemaps-entries submitted for this site, or included in the sitemap index file (if sitemapIndex is specified in the request).

SitemapMethods
A builder providing access to all methods supported on sitemap resources. It is not used directly, but through the Webmasters hub.

SitemapSubmitCall
Submits a sitemap for a site.

SitemapsListResponse
List of sitemaps.

SitesListResponse
List of sites with access level information.

UrlCrawlErrorCount
An entry in a URL crawl errors time series.

UrlCrawlErrorCountsPerType
Number of errors per day for a specific error type (defined by platform and category).

UrlCrawlErrorsCountsQueryResponse
A time series of the number of URL crawl errors per error category and platform.

UrlCrawlErrorsSample
Contains information about specific crawl errors.

UrlCrawlErrorsSamplesListResponse
List of crawl error samples.

UrlSampleDetails
Additional details about the URL, set only when calling get().

UrlcrawlerrorscountMethods
A builder providing access to all methods supported on urlcrawlerrorscount resources. It is not used directly, but through the Webmasters hub.

UrlcrawlerrorscountQueryCall
Retrieves a time series of the number of URL crawl errors per error category and platform.

UrlcrawlerrorssampleGetCall
Retrieves details about crawl errors for a site's sample URL.

UrlcrawlerrorssampleListCall
Lists a site's sample URLs for the specified crawl error category and platform.

UrlcrawlerrorssampleMarkAsFixedCall
Marks the provided site's sample URL as fixed, and removes it from the samples list.

UrlcrawlerrorssampleMethods
A builder providing access to all methods supported on urlcrawlerrorssample resources. It is not used directly, but through the Webmasters hub.

Webmasters
Central instance to access all Webmasters related resource activities

WmxSite
Contains permission level information about a Search Console site. For more information, see Permissions in Search Console.

WmxSitemap
Contains detailed information about a specific URL submitted as a sitemap.

WmxSitemapContent
Information about the various content types in the sitemap.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.