Struct google_webmasters3::SearchAnalyticsQueryRequest

pub struct SearchAnalyticsQueryRequest {
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub dimensions: Option<Vec<String>>,
    pub search_type: Option<String>,
    pub dimension_filter_groups: Option<Vec<ApiDimensionFilterGroup>>,
    pub aggregation_type: Option<String>,
    pub row_limit: Option<i32>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query searchanalytics (request)

Fields

start_date

[Required] Start date of the requested date range, in YYYY-MM-DD format, in PST time (UTC - 8:00). Must be less than or equal to the end date. This value is included in the range.

end_date

[Required] End date of the requested date range, in YYYY-MM-DD format, in PST (UTC - 8:00). Must be greater than or equal to the start date. This value is included in the range.

dimensions

[Optional] Zero or more dimensions to group results by. Dimensions are the group-by values in the Search Analytics page. Dimensions are combined to create a unique row key for each row. Results are grouped in the order that you supply these dimensions.

search_type

[Optional; Default is "web"] The search type to filter for.

dimension_filter_groups

[Optional] Zero or more filters to apply to the dimension grouping values; for example, 'query contains "buy"' to see only data where the query string contains the substring "buy" (not case-sensitive). You can filter by a dimension without grouping by it.

aggregation_type

[Optional; Default is "auto"] How data is aggregated. If aggregated by property, all data for the same property is aggregated; if aggregated by page, all data is aggregated by canonical URI. If you filter or group by page, choose AUTO; otherwise you can aggregate either by property or by page, depending on how you want your data calculated; see the help documentation to learn how data is calculated differently by site versus by page.

Note: If you group or filter by page, you cannot aggregate by property.

If you specify any value other than AUTO, the aggregation type in the result will match the requested type, or if you request an invalid type, you will get an error. The API will never change your aggregation type if the requested type is invalid.

row_limit

[Optional; Default is 1000] The maximum number of rows to return. Must be a number from 1 to 5,000 (inclusive).