Struct google_webmasters3::UrlcrawlerrorssampleMethods

pub struct UrlcrawlerrorssampleMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on urlcrawlerrorssample resources. It is not used directly, but through the Webmasters hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_webmasters3 as webmasters3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use webmasters3::Webmasters;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Webmasters::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `list(...)` and `mark_as_fixed(...)`
// to build up your call.
let rb = hub.urlcrawlerrorssamples();

Methods

impl<'a, C, A> UrlcrawlerrorssampleMethods<'a, C, A>

fn get(&self, site_url: &str, url: &str, category: &str, platform: &str) -> UrlcrawlerrorssampleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves details about crawl errors for a site's sample URL.

Arguments

siteUrl - The site's URL, including protocol. For example: http://www.example.com/

url - The relative path (without the site) of the sample URL. It must be one of the URLs returned by list(). For example, for the URL https://www.example.com/pagename on the site https://www.example.com/, the url value is pagename

category - The crawl error category. For example: authPermissions

platform - The user agent type (platform) that made the request. For example: web

fn list(&self, site_url: &str, category: &str, platform: &str) -> UrlcrawlerrorssampleListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists a site's sample URLs for the specified crawl error category and platform.

Arguments

siteUrl - The site's URL, including protocol. For example: http://www.example.com/

category - The crawl error category. For example: authPermissions

platform - The user agent type (platform) that made the request. For example: web

fn mark_as_fixed(&self, site_url: &str, url: &str, category: &str, platform: &str) -> UrlcrawlerrorssampleMarkAsFixedCall<'a, C, A>

Create a builder to help you perform the following task:

Marks the provided site's sample URL as fixed, and removes it from the samples list.

Arguments

siteUrl - The site's URL, including protocol. For example: http://www.example.com/

url - The relative path (without the site) of the sample URL. It must be one of the URLs returned by list(). For example, for the URL https://www.example.com/pagename on the site https://www.example.com/, the url value is pagename

category - The crawl error category. For example: authPermissions

platform - The user agent type (platform) that made the request. For example: web