Struct google_webmasters3::WmxSitemap

pub struct WmxSitemap {
    pub errors: Option<String>,
    pub warnings: Option<String>,
    pub is_pending: Option<bool>,
    pub last_submitted: Option<String>,
    pub is_sitemaps_index: Option<bool>,
    pub last_downloaded: Option<String>,
    pub path: Option<String>,
    pub type_: Option<String>,
    pub contents: Option<Vec<WmxSitemapContent>>,
}

Contains detailed information about a specific URL submitted as a sitemap.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get sitemaps (response)

Fields

errors

Number of errors in the sitemap. These are issues with the sitemap itself that need to be fixed before it can be processed correctly.

warnings

Number of warnings for the sitemap. These are generally non-critical issues with URLs in the sitemaps.

is_pending

If true, the sitemap has not been processed.

last_submitted

Date & time in which this sitemap was submitted. Date format is in RFC 3339 format (yyyy-mm-dd).

is_sitemaps_index

If true, the sitemap is a collection of sitemaps.

last_downloaded

Date & time in which this sitemap was last downloaded. Date format is in RFC 3339 format (yyyy-mm-dd).

path

The url of the sitemap.

type_

The type of the sitemap. For example: rssFeed.

contents

The various content types in the sitemap.